storing data locally as opposed to a warehouse and split the script to make them less bulky
