+++ Database/sqlwrapper.py
@@ -1,7 +1,7 @@
-
 import psycopg2
 import psycopg2.extras
 import os
+from dotenv import load_dotenv
 
 load_dotenv(override=True, verbose=True)
